src/operator/requirements.txt
kubernetes
pykube-ng
kopf
asyncio
shortuuid